backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
prophet
google-genai
firebase-admin
requests
pymongo